feat: scientific session supervisor

+++ lib/src/data/models/assessment/list_scientific_assignment.dart
@@ -18,7 +18,7 @@ class ListScientificAssignment {
   @JsonKey(name: "scores")
   List<Score>? scores;
   @JsonKey(name: "grade")
-  int? grade;
+  double? grade;
 
   ListScientificAssignment({
     this.id,
